backend/requirements.txt
# ── Núcleo ─────────────────────────────────────────────────────────
flask>=3.0.0
flask-cors>=4.0.0

# ── Base de datos ──────────────────────────────────────────────────
psycopg2-binary>=2.9.0

# ── Variables de entorno ───────────────────────────────────────────
python-dotenv>=1.0.0

# ── Autenticación ──────────────────────────────────────────────────
PyJWT>=2.8.0
bcrypt>=4.1.0

# ── Servidor de producción ─────────────────────────────────────────
gunicorn>=22.0.0
